On demand injection of Tracer Header
Currently the Sling Tracer headers are being sent for all active tabs as the listener is global. Changing that to now only do that for requests originating from those tabs where the AEM Panel is active in DevTools.

Upon AEM Panel activation the tab would be registered for header injection. This would ensure that plugin does not interfere with global request processing and headers only get injected for those pages which the developer wants to debug.

Later we can make it more precise if AEM Panel support option of explicit activation of tracing.